requirements.txt
google-api-python-client
google-auth-oauthlib
gspread
beautifulsoup4
anthropic
supabase
python-dotenv
streamlit
pandas
